+++ .gitignore
@@ -1,6 +1,7 @@
 # Custom Settings
 test/
 /app/config/
+checkpoints/
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
